requirements.txt
asgiref==3.7.2
azure-core==1.29.7
azure-storage-blob==12.19.0
beautifulsoup4==4.12.3
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
crispy-bootstrap5==2023.10
cryptography==42.0.0
Django==4.2.9
django-bootstrap5==23.4
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-storages==1.14.2
djangorestframework==3.14.0
et-xmlfile==1.1.0
gunicorn==22.0.0
idna==3.6
isodate==0.6.1
numpy==1.26.4
openpyxl==3.1.2
packaging==24.0
pandas==2.2.0
pillow==10.2.0
psycopg2-binary==2.9.9
pybase64==1.3.2
pycparser==2.21
pyOpenSSL==24.1.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
requests==2.31.0
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2024.1
urllib3==1.26.15
Wikipedia-API==0.6.0
